flask-project/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
PyJWT==2.8.0
pg8000==1.30.1
Flask-SocketIO
requests
python-dotenv
eventlet
